Cleanup

+++ src/main/java/gcewing/architecture/BaseItemBlock.java
@@ -6,16 +6,19 @@
 
 package gcewing.architecture;
 
-import static gcewing.architecture.BaseBlockUtils.*;
-import static gcewing.architecture.BaseUtils.*;
+import static gcewing.architecture.BaseBlockUtils.getBlockStateFromMeta;
+import static gcewing.architecture.BaseBlockUtils.getMetaFromBlockState;
+import static gcewing.architecture.BaseUtils.facings;
 
-import net.minecraft.block.*;
+import net.minecraft.block.Block;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.*;
+import net.minecraft.item.ItemBlock;
+import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.world.World;
 
-import gcewing.architecture.BaseMod.*;
+import gcewing.architecture.BaseMod.IItem;
+import gcewing.architecture.BaseMod.ModelSpec;
 
 public class BaseItemBlock extends ItemBlock implements IItem {
 
@@ -58,9 +61,9 @@ public class BaseItemBlock extends ItemBlock implements IItem {
             int j1 = baseBlock.getMetaFromState(state);
             if (placeBlockAt(stack, player, world, x, y, z, side, hitX, hitY, hitZ, j1)) {
                 world.playSoundEffect(
-                        (double) ((float) x + 0.5F),
-                        (double) ((float) y + 0.5F),
-                        (double) ((float) z + 0.5F),
+                        (float) x + 0.5F,
+                        (float) y + 0.5F,
+                        (float) z + 0.5F,
                         this.field_150939_a.stepSound.func_150496_b(),
                         (this.field_150939_a.stepSound.getVolume() + 1.0F) / 2.0F,
                         this.field_150939_a.stepSound.getPitch() * 0.8F);
