+++ src/main/java/gcewing/architecture/InventoryHelper.java
@@ -11,7 +11,8 @@ import java.util.Random;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.item.EntityItem;
 import net.minecraft.inventory.IInventory;
-import net.minecraft.item.*;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.world.World;
 
@@ -20,7 +21,7 @@ public class InventoryHelper {
     private static final Random RANDOM = new Random();
 
     public static void dropInventoryItems(World worldIn, BlockPos pos, IInventory p_180175_2_) {
-        func_180174_a(worldIn, (double) pos.getX(), (double) pos.getY(), (double) pos.getZ(), p_180175_2_);
+        func_180174_a(worldIn, pos.getX(), pos.getY(), pos.getZ(), p_180175_2_);
     }
 
     public static void func_180176_a(World worldIn, Entity p_180176_1_, IInventory p_180176_2_) {
